feat: extract workspace_symbols to features layer
- Create features/workspace_symbols.rs with compute_workspace_symbols()
  taking SymbolIndex + SearchAccess bounds
- WorkspaceSymbolQuery, collect_index_symbols, rg_symbol_search,
  symbol_information helpers all move out of handlers.rs
- symbol_impl becomes a 5-line adapter
- traits_impl: for_each_indexed_file delegates to the concrete method
  (was duplicating the loop body)
- Removes #[allow(dead_code)] on for_each_indexed_file trait method

handlers.rs is now 128 lines (was 258 before this session)
