+++ src/features/mod.rs
@@ -26,3 +26,4 @@ pub(crate) mod symbols;
 pub(crate) mod text_utils;
 pub(crate) mod traits;
 mod traits_impl;
+pub(crate) mod workspace_symbols;
